+++ Task 4/Features/Common/Extension/UIImageViewExt.swift
@@ -8,15 +8,18 @@
 import UIKit
 
 extension UIImageView {
-    // MARK: - Add round corner
+    
+    // Add round corner
     func roundCorner(cornerRadius : CGFloat) {
+        
         layer.cornerRadius = cornerRadius
         clipsToBounds = true
         contentMode = .scaleAspectFill
     }
     
-    // MARK: - Add shadow
+    // Add shadow
     func addShadow(containerView: UIView, color : CGColor, shadowOpacity: Float, shadowRadius: Double, cornerRadius: CGFloat) {
+        
         containerView.clipsToBounds = false
         containerView.layer.shadowColor = color
         containerView.layer.shadowOpacity = shadowOpacity
@@ -24,6 +27,7 @@ extension UIImageView {
         containerView.layer.shadowRadius = shadowRadius
         containerView.backgroundColor = .clear
         containerView.layer.shadowPath = UIBezierPath(roundedRect: containerView.bounds, cornerRadius: cornerRadius).cgPath
+        
         clipsToBounds = true
         layer.cornerRadius = cornerRadius
         contentMode = .scaleAspectFill
