+++ Task 4/Features/Common/Extension/UISegmentedControlExt.swift
@@ -9,7 +9,7 @@ import UIKit
 
 extension UISegmentedControl {
 
-    // MARK: - Set segmentedControl title
+    // Set segmentedControl title
     func addTitle (titles: [String]) {
         for i in 0 ..< titles.count {
             self.setTitle(titles[i], forSegmentAt: i)
