+++ Task 4/Features/Common/Extension/UITabBarItem.swift
@@ -9,7 +9,7 @@ import UIKit
 
 extension UITabBarItem {
     
-    // MARK: - Set tabBar image and title 
+    // Set tabBar image and title 
     func customizeTabBar(title: String!, image: UIImage!, selectedImage: UIImage!) {
         self.title = title
         self.image = image
