+++ Task 4/Features/Movies/Views/MoviesTVC.swift
@@ -12,22 +12,17 @@ import UICircularProgressRing
 class MoviesTVC: UITableViewCell {
 
     @IBOutlet weak var backView: UIView!
-    @IBOutlet weak var gradientView: UIView!
-    
     @IBOutlet weak var imgPoster: UIImageView!
     @IBOutlet weak var lblTitle: UILabel!
     @IBOutlet weak var lblReleaseDate: UILabel!
     @IBOutlet weak var lblOverview: UILabel!
     @IBOutlet weak var lblVoteAverage: UILabel!
     @IBOutlet weak var movieRate: UICircularProgressRing!
-    
     @IBOutlet weak var btnGoToDetails: UIButton!
+    
     override func awakeFromNib() {
         super.awakeFromNib()
         
-//        gradientView.addGradient()
-//        backView.backgroundColor = .clear
-        gradientView.backgroundColor = .clear
         imgPoster.roundCorner(cornerRadius: 15)
     }
 
@@ -37,13 +32,14 @@ class MoviesTVC: UITableViewCell {
     }
     
     func setupCell(title: String?, releaseDate: String?, overview: String?, rate: Double?,
-                   voteAverage: Double?, imageUrl: URL?) {
+                   voteAverage: Double?, imageUrl: String?) {
+        
         lblTitle.text = title ?? ""
         lblReleaseDate.text = releaseDate ?? ""
         lblOverview.text = overview ?? ""
         movieRate.value = rate ?? 0
         lblVoteAverage.text = String(describing: (voteAverage ?? 0))
-        imgPoster.kf.setImage(with: imageUrl, options: [.cacheOriginalImage])
+        imgPoster.kf.setImage(with: URL(string: imageUrl ?? ""), options: [.cacheOriginalImage])
         
     }
