+++ Task 4/Features/People/Controllers/ActorVC.swift
@@ -40,32 +40,10 @@ class ActorVC: UIViewController {
             options: [.cacheOriginalImage] )
     }
 
-    @IBAction func btnBack(_ sender: Any) {
+    @IBAction func btnBackTapped(_ sender: Any) {
         self.dismiss(animated: true)
     }
     
 
 }
 
-extension ActorVC: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        guard let count = actor?.knownFor?.count else {return 0}
-        return count
-    }
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        if let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "knownForCell", for: indexPath) as? knownForCVC {
-            let actorInfo = self.actor?.knownFor?[indexPath.row]
-            cell.imgMoviePoster.kf.setImage(with: URL(string: Api.baseImageUrl+(actorInfo?.posterPath ?? "")), options: [.cacheOriginalImage])
-            cell.imgMoviePoster.roundCorner(cornerRadius: 15)
-            return cell
-        } else {
-            return UICollectionViewCell()
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: (collectionView.frame.width) / 2.5, height: collectionView.frame.height)
-
-    }
-}
