+++ Task 4/Model/SegmentState.swift
@@ -1,0 +1,14 @@
+//
+//  SegmentState.swift
+//  Task 4
+//
+//  Created by ME-MAC on 1/5/23.
+//
+
+import Foundation
+
+// Store page num and movie list for each segment
+struct MovieList {
+    var pageNumber: Int = 0
+    var movie: [Movies] = [Movies]()
+}
